voicetalk-src/requirements.txt
Flask~=1.1.2
Flask-Login~=0.5.0
alembic~=1.4.2
SQLAlchemy~=1.3.18
PyMySQL~=0.10.0
requests~=2.24.0
uWSGI~=2.0.19.1
argon2-cffi~=20.1.0
